fix: bug where version check was failing
fix: ignore rescan response
chore: enable 2fa for app store
fix: filter button disapearing when filtering for non existant things

+++ FullyNoded/View Controllers/Wallets/SeedDisplayerViewController.swift
@@ -18,7 +18,7 @@ class SeedDisplayerViewController: UIViewController, UINavigationControllerDeleg
     var name = ""
     var coinType = "0"
     var blockheight:Int64!
-    var version:Int = 0
+    var version:Double = 0.0
     
     override func viewDidLoad() {
         super.viewDidLoad()
